Refactor ChatView components environments initialization
