+++ GliaWidgets/Sources/Component/Connect/ConnectView.Environment.swift
@@ -1,0 +1,23 @@
+import Foundation
+
+extension ConnectView {
+    struct Environment {
+        var data: FoundationBased.Data
+        var uuid: () -> UUID
+        var gcd: GCD
+        var imageViewCache: ImageView.Cache
+        var timerProviding: FoundationBased.Timer.Providing
+    }
+}
+
+extension ConnectView.Environment {
+    static func create(with environment: ChatView.Environment) -> Self {
+        .init(
+            data: environment.data,
+            uuid: environment.uuid,
+            gcd: environment.gcd,
+            imageViewCache: environment.imageViewCache,
+            timerProviding: environment.timerProviding
+        )
+    }
+}
