+++ GliaWidgets/Sources/Component/Connect/ConnectView.swift
@@ -198,16 +198,6 @@ private extension ConnectView {
     }
 }
 
-extension ConnectView {
-    struct Environment {
-        var data: FoundationBased.Data
-        var uuid: () -> UUID
-        var gcd: GCD
-        var imageViewCache: ImageView.Cache
-        var timerProviding: FoundationBased.Timer.Providing
-    }
-}
-
 extension ConnectView.State {
     var chatContentSpacing: CGFloat {
         switch self {
