+++ GliaWidgets/Sources/Component/UnreadMessageIndicator/UnreadMessageIndicatorView.Environment.swift
@@ -1,0 +1,21 @@
+import Foundation
+
+extension UnreadMessageIndicatorView {
+    struct Environment {
+        var data: FoundationBased.Data
+        var uuid: () -> UUID
+        var gcd: GCD
+        var imageViewCache: ImageView.Cache
+    }
+}
+
+extension UnreadMessageIndicatorView.Environment {
+    static func create(with environment: ChatView.Environment) -> Self {
+        .init(
+            data: environment.data,
+            uuid: environment.uuid,
+            gcd: environment.gcd,
+            imageViewCache: environment.imageViewCache
+        )
+    }
+}
