+++ GliaWidgets/Sources/View/Chat/Entry/ChatMessageEntryView.Environment.swift
@@ -1,0 +1,19 @@
+import Foundation
+
+extension ChatMessageEntryView {
+    struct Environment {
+        var gcd: GCD
+        var uiApplication: UIKitBased.UIApplication
+        var uiScreen: UIKitBased.UIScreen
+    }
+}
+
+extension ChatMessageEntryView.Environment {
+    static func create(with environment: ChatView.Environment) -> Self {
+        .init(
+            gcd: environment.gcd,
+            uiApplication: environment.uiApplication,
+            uiScreen: environment.uiScreen
+        )
+    }
+}
