+++ GliaWidgets/Sources/View/Chat/GVA/PersistentButton/GvaPersistentButtonView.Environment.swift
@@ -1,0 +1,23 @@
+import Foundation
+
+extension GvaPersistentButtonView {
+    struct Environment {
+        var data: FoundationBased.Data
+        var uuid: () -> UUID
+        var gcd: GCD
+        var imageViewCache: ImageView.Cache
+        var uiScreen: UIKitBased.UIScreen
+    }
+}
+
+extension GvaPersistentButtonView.Environment {
+    static func create(with environment: ChatView.Environment) -> Self {
+        .init(
+            data: environment.data,
+            uuid: environment.uuid,
+            gcd: environment.gcd,
+            imageViewCache: environment.imageViewCache,
+            uiScreen: environment.uiScreen
+        )
+    }
+}
