+++ GliaWidgets/Sources/View/Chat/Message/ChatMessageView.Environment.swift
@@ -1,0 +1,13 @@
+import Foundation
+
+extension ChatMessageView {
+    struct Environment {
+        var uiScreen: UIKitBased.UIScreen
+    }
+}
+
+extension ChatMessageView.Environment {
+    static func create(with environment: ChatView.Environment) -> Self {
+        .init(uiScreen: environment.uiScreen)
+    }
+}
