+++ GliaWidgets/Sources/View/Chat/Message/SystemMessageView.Environment.swift
@@ -1,0 +1,13 @@
+import Foundation
+
+extension SystemMessageView {
+    struct Environment {
+        var uiScreen: UIKitBased.UIScreen
+    }
+}
+
+extension SystemMessageView.Environment {
+    static func create(with environment: ChatView.Environment) -> Self {
+        .init(uiScreen: environment.uiScreen)
+    }
+}
